Mon, Aug 21, 2023, 10:50 PM +04:00

+++ readability/readability.c
@@ -78,8 +78,8 @@ int main(void)
     //Using Coleman-Liau formula to calculate the grade of the text
     int calculation(int letters, int words, int sentences)
     {
-        float L = (letters/words)*100;
-        float S = (sentences/words)*100;
+        float L = ((float)letters/words)*100;
+        float S = ((float)sentences/words)*100;
         float index = (0.0588 * L) - (0.296 * S) - 15.8;
 
         int X = round(index);
